Add distance modifier in ellipse fit

+++ reborn/viewers/qtviews/plugins/correct_polarization.py
@@ -15,7 +15,7 @@
 
 def plugin(padview):
     r""" Plugin for PADView. Divides out the beam polarization factor. """
-    data = padview.dataframe.get_processed_data_flat()
+    data = padview.dataframe.get_processed_data_flat().astype(float)
     beam = padview.dataframe.get_beam()
     geom = padview.dataframe.get_pad_geometry()
     data /= geom.polarization_factors(beam=beam)
